Made it so food objects can now be displayed along with a running total of calorie count
Made a nutrition page which will display all of the food objects within the 'foods' table. Also included a daily calorie count which will change based on if you have gone over your calorie limit (calculated from the food objects). Basic functionality is there, we now need to tie this to a specific user if possible.